dev_requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file

debugpy==1.2.1
black==20.8b1
ipykernel==5.5.0
ipython==7.20.0
ipython-genutils==0.2.0
jupyter-client==6.1.11
jupyter-core==4.7.1
ptyprocess==0.7.0
Pygments==2.8.0
pylint==2.7.1
pylint-flask-sqlalchemy==0.2.0
python-dateutil==2.8.1
python-editor==1.0.4
pytz==2021.1
pyzmq==22.0.3
regex==2020.11.13
six==1.15.0
typed-ast==1.4.2
typing-extensions==3.7.4.3
yapf==0.30.0
